+++ experimental/m/meitei_legacy/source/help/meitei_legacy.php
@@ -1,0 +1,19 @@
+<?php
+  $pagename = 'Meitei (Legacy) Keyboard Help';
+  $pagetitle = 'Meitei (Legacy) Keyboard Help';
+  require_once('header.php');
+?>
+<p>
+This is a keyboard for entering Unicode data in the Meitei script.
+It follows the keystrokes of the legacy font Ratha99, but produces Unicode data.
+</p>
+
+<h2>Keyboard Layout</h2>
+
+<img src="meitei_legacy.png" alt="Meitei keyboard layout" style="width:600px;">
+
+<h2 Usage Notes</h2>
+<p>
+Type "/" before a key to get the original English character.
+
+<p>© 2019 SIL International</p>