+++ experimental/w/wakhi_cyrillic/source/help/wakhi_cyrillic.php
@@ -1,0 +1,32 @@
+<?php
+$pagename = 'Wakhi Cyrillic Keyboard Help';
+$pagetitle = $pagename;
+// Header
+require_once('header.php');
+?>
+
+<p>
+Wakhi Cyrillic keyboard.
+</p>
+<p>
+For the letters combined with diacritic symbols, first type the letter, then type the diacritic symbol. The diacritic symbols are on the top row under the keys 1, 2 and 3.
+
+<p>
+If you have problem using the Wakhi Cyrillic keyboard please contact: writing.xikwor@gmail.com
+</p>
+
+<p> См. здесь <a href="file:wakhi_cyrillic_help.pdf">здесь</a> для инструкций на русском языке.</p>
+<h1>Keyboard Layout</h1>
+
+<h1>Keyboard Layout</h1>
+
+<h2>Desktop Keyboard Layout</h2>
+<div id='osk' data-states='default shift ctrl'>
+</div>
+
+<h2>Mobile Keyboard Layout</h2>
+<p>Due to the size and number of keys, some characters are hidden in the long press.
+Press and hold on the key with a little dot on the top right to reveal them.</p>
+
+<div id='osk-phone' data-states='default shift numeric'>
+</div>