+++ legacy/p/pothana/source/help/pothana.php
@@ -1,0 +1,11 @@
+<?php 
+  $pagename = 'Pothana Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+
+<p>
+    Documentation for this keyboard is available here: <a href="manual.pdf">manual.pdf</a>
+</p>