+++ release/basic/basic_kbdinkan/source/help/basic_kbdinkan.php
@@ -1,0 +1,27 @@
+<?php 
+  $pagename = 'Kannada Basic Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+<p>
+    Kannada Basic 1.1 generated from template.
+</p>
+
+<h1>Keyboard Layout</h1>
+
+<h2>Desktop Keyboard Layout</h2>
+<div id='osk' data-states='default shift rightalt rightalt-shift leftctrl-shift'>
+</div>
+
+<h2>Mobile/Tablet Keyboard Layout</h2>
+
+<div id='osk-tablet' data-states='default shift numeric'>
+</div>
+
+
+<p>(c) 2009-2019 SIL International</p>
+
+</body>
+</html>