+++ release/basic/basic_kbdmlt47/source/help/basic_kbdmlt47.php
@@ -1,0 +1,26 @@
+<?php 
+  $pagename = 'Maltese 47-Key Basic Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+<p>
+    Maltese 47-Key Basic 1.1 generated from template.
+</p>
+<p>This keyboard defines some keys that are set by default as "hotkeys" when Keyman is installed. 
+To use this keyboard, you will want to go to <b>Configuration.../Hotkeys</b>, 
+select the hotkey name, and either change the hotkey or turn it off. 
+Currently 
+"Switch Keyman Desktop Off" uses Shift+Alt+O.</p>
+
+<h1>Keyboard Layout</h1>
+
+<h2>Desktop Keyboard Layout</h2>
+<div id='osk' data-states='default shift rightalt rightalt-shift'>
+</div>
+
+<h2>Mobile/Tablet Keyboard Layout</h2>
+
+<div id='osk-tablet' data-states='default shift rightalt'>
+</div>