+++ release/e/english_shavian_igc/source/help/english_shavian_igc.php
@@ -1,0 +1,52 @@
+﻿<?php 
+  $pagename = 'Shavian IGC Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+<p>
+
+The Shavian alphabet is an alternative, phonemic way to write English. The alphabet takes its name from Irish writer Bernard Shaw who inspired and posthumously funded it. Shaw believed there should be a simpler way to write English. His wish was fulfilled in 1962 following a worldwide competition. The winner of that competition, Kingsley Read, developed an alphabet designed especially for English, with a unique letter for each of the 40 or so basic sounds we all use today.
+
+</p>
+
+<img src="shavian_alphabet_reading_key.png" width=100%>
+
+<p>
+
+For further information about the Shavian alphabet visit <a href="https://www.shavian.info">Shavian.info</a>.
+
+</p>
+
+<h1>Keyboard Layout</h1>
+
+<h2>Desktop and onscreen layouts</h2>
+
+<p>The desktop and onscreen layouts are identical, and are based on the <a href="https://oztypewriter.blogspot.com/2014/04/history-of-shavian-keyboard-imperial.html">Shavian Imperial Good Companion</a> typewriter layout, used by by Kingsley Read in the 1960s. Shavian IGC is named in honour of the Shavian Imperial Good Companion typewriter.</p>
+<p>The compound letters (𐑸, 𐑹, 𐑺, 𐑻, 𐑼, 𐑽, & 𐑿), and a range of symbols, are entered by holding down SHIFT.</p>
+<p>Numbers may be entered by holding down ALT.</p>
+
+<div id='osk' data-states='default shift alt'>
+</div>
+
+<H2>Phone and tablet layouts</H2>
+
+<p>The layouts for phone and tablet are identical to each other, and are based on the desktop layout with some tweaks to avoid the keys becoming too small.</p>
+
+<p>The compound letters (and 𐑠) may be found by holding down (long hold) one of the keys with the dot in the top right. Holding down:<p>
+<ul>
+<li>𐑭 will display 𐑸</li>
+<li>𐑷 will display 𐑹</li>
+<li>𐑵 will display 𐑿</li>
+<li>𐑱 will display 𐑺</li>
+<li>𐑳 will display 𐑻</li>
+<li>𐑦 will display 𐑾 and 𐑽</li>
+<li>𐑩 will display 𐑼</li>
+<li>𐑖 will display 𐑠</li>
+</ul>
+
+<p>Pressing the 123 key will bring up the numbers and a range of symbols, arranged in a similar layout to the standard iOS keyboard.</p>
+
+<div id='osk-tablet' data-states='default alt'>
+</div>