+++ release/m/malar_malayalam/source/help/malar_malayalam.php
@@ -1,0 +1,16 @@
+<?php
+  $pagename = 'Malar Malayalam Keyboard Help';
+  $pagetitle = 'Malar Malayalam Keyboard Help';
+  $pagestyle = <<<END
+  
+     
+		 
+END;
+  require_once('header.php');
+?>
+
+<p><b>Malar Malayalam</b> Keyboard is a transliteration based Malayalam input method. This keyboard supports all Malayalam characters encoded up to the Unicode Standard-Version 12.1. Malayalam Unicode fonts under the SIL Open Font License are available at <a href="https://smc.org.in/fonts/" target="_blank">https://smc.org.in/fonts/</a> and <a href="http://rachana.org.in/" target="_blank">http://rachana.org.in/</a>. Refer the below chart for key mapping. PDF version of the chart is available <a href="./malar_malayalam.pdf" target="_blank">here</a>.</p>
+
+<img src="./malar_malayalam.jpg" alt="Malar Malayalam Key Mappings" style="display: block; margin-left: auto; margin-right: auto; max-width:100%; width:auto; height:auto">
+
+<p><a href="https://forms.gle/DcJZZBHH4vwyoLoz7" target="_blank">Click here to submit a feedback about <b>Malar Malayalam</b> Keyboard.</a></p>