+++ release/packages/vm_tamil/source/help/vm_tamil.php
@@ -1,0 +1,19 @@
+<?php
+  $pagename = 'Tamil (Visual Media) Keyboard Package Help';
+  $pagetitle = $pagename;
+  require_once('header.php');
+?>
+
+<h1>Tamil (Visual Media) Keyboard Package</h1>
+
+<p style='margin:0px'>&#169; 2008-2020 Visual Media and SIL International</p>
+
+<p>
+This package contains the Visual Media Tamil Modular and Tamil Typewriter keyboards for Keyman Desktop.  Information on using each of these keyboards is on separate pages.
+</p>
+
+  <section id='related'>
+	<h2>Keyboard Layouts</h2>
+<ul><a href="http://help.keyman.com/keyboard/vm_tamil_modular">Tamil Modular Keyboard</a> </ul>
+<ul><a href="https://help.keyman.com/keyboard/vm_tamil_typewriter">Tamil Typewriter Keyboard</a> </ul>
+  </section>