+++ release/packages/fv_all/HISTORY.md
@@ -1,6 +1,11 @@
 # fv_all Keyboard Package
 
-## 12.0 (14 Jun 2019)
+## 12.2 (29 Apr 2020)
+* Add touch layout keyboards and DejaVuSans.ttf font
 
+## 12.1 (3 Apr 2020)
+* Added fv_haisla
+
+## 12.0 (14 Jun 2019)
 * Initial version for distribution with FV 12.0