+++ release/s/soqotri_arabic/HISTORY.md
@@ -1,7 +1,10 @@
 Soqotri Keyboard Change History
 =======================
 
+1.1 (1 May 2020)
+-----------------
+* Bug fix in mobile layout to access shift layer
+
 1.0 (27 Jun 2019)
 -----------------
-
 * Created by SIL